+++ tests/unit/CommandInterpreterTest.php
@@ -713,9 +713,9 @@ class CommandInterpreterTest extends TestCase {
 		$interpreter = new class() extends Command_Interpreter_Node {
 			/** @var string[] */
 			public array $dropped = [];
-			public function drop_message( array $message, string $error, string $node = '' ): void {
+			public function drop_message( array $message, string $error ): void {
 				$this->dropped[] = $error;
-				parent::drop_message( $message, $error, $node );
+				parent::drop_message( $message, $error );
 			}
 		};
 		$interpreter->name( '_command_interpreter' );
@@ -742,9 +742,9 @@ class CommandInterpreterTest extends TestCase {
 		$interpreter = new class() extends Command_Interpreter_Node {
 			/** @var string[] */
 			public array $dropped = [];
-			public function drop_message( array $message, string $error, string $node = '' ): void {
+			public function drop_message( array $message, string $error ): void {
 				$this->dropped[] = $error;
-				parent::drop_message( $message, $error, $node );
+				parent::drop_message( $message, $error );
 			}
 		};
 		$interpreter->name( '_command_interpreter' );